requirements_enhanced.txt
# Enhanced Prescription System Requirements
fpdf2>=2.7.9
pandas>=1.3.0
cryptography>=3.4.8